+++ examples/resources/nxos_feature_netflow/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_netflow.example "sys/fm/netflow"