+++ examples/resources/nxos_feature_netflow/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_netflow" "example" {
+  admin_state = "enabled"
+}